pyproject.toml
[tool.poetry]
name = "nonebot_plugin_wordle"
version = "0.2.0"
description = "Nonebot2 wordle猜单词插件"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-wordle"
repository = "https://github.com/noneplugin/nonebot-plugin-wordle"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = { version = "^2.0.0-rc.1", extras = ["fastapi"] }
nonebot-adapter-onebot = "^2.2.0"
Pillow = "^9.0.0"
pyspellchecker = "^0.7.0"

[tool.poetry.group.dev.dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
